added the main.java file back
